fix: 코드 파일 분리 처리
